# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TestAbsRequest": {
        "properties": {
          "token": {
            "default": "",
            "title": "Token",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "TestAbsRequest",
        "type": "object"
      },
      "TestHardcoverRequest": {
        "properties": {
          "token": {
            "default": "",
            "title": "Token",
            "type": "string"
          }
        },
        "title": "TestHardcoverRequest",
        "type": "object"
      },
      "TestRemoteRequest": {
        "properties": {
          "key": {
            "default": "",
            "title": "Key",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "TestRemoteRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Synchronize your reading position between ebooks and audiobooks. Seamlessly switch between reading and listening.",
    "title": "Tandem",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings/": {
      "get": {
        "description": "System settings: everything for an admin, an allow-list for everyone else.",
        "operationId": "get_settings_api_settings__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Settings Api Settings  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "Update system settings.",
        "operationId": "update_settings_api_settings__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "New Settings",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Settings Api Settings  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/test-abs": {
      "post": {
        "description": "Test the connection to an Audiobookshelf server.",
        "operationId": "test_abs_connection_api_settings_test_abs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestAbsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Test Abs Connection",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/test-hardcover": {
      "post": {
        "description": "Validate a Hardcover API token by running the trivial `me` query.\nSame placeholder/stored-credential fallback semantics as test-abs.",
        "operationId": "test_hardcover_connection_api_settings_test_hardcover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestHardcoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Test Hardcover Connection",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/test-remote": {
      "post": {
        "description": "Test the connection to a remote transcription server from the backend.\nThis avoids CORS and VPN routing issues where the frontend browser\ncannot reach a local LAN IP like the Jetson directly.",
        "operationId": "test_remote_connection_api_settings_test_remote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRemoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Test Remote Connection",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/transcription-remote-key/generate": {
      "post": {
        "description": "Generate and persist a new shared secret for the Jetson transcription\nserver. Returned once, in plaintext, so the admin can copy it into the\nJetson's TRANSCRIPTION_API_KEY — it's stored encrypted from here on and\nGET /api/settings only ever returns the masked placeholder for it.",
        "operationId": "generate_transcription_remote_key_api_settings_transcription_remote_key_generate_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Generate Transcription Remote Key",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}